node starts polling loop for visa service (#201)
This adds a visa service "handler" to the node main loop. There is support for registering with the visa service, polling, and de-registering. No support for the request-visa call or the authenticate-connect call.

This is implemented as a placeholder since normally this would be kicked off until such a time as the visa service adapter connects. So for now, you can pass a flag to the node when you launch it to force it to connect to a listening visa service. Hitting CTRL-C will stop the node, and deregister from the visa service.