boxed packet buffers (#521)
In preparation for removing 'pktbuf from Assembly, this commit makes
buffer_stack own the buffers it manages.  It owns them as Boxes, so there is
no performance hit (i.e.  no reference counting).  The difference though is
that this prevents us from allocating a single contiguous slab of memory and
divvying it up in order to avoid malloc header overhead.  However it's
possible we can leverage the yoke crate to give us this capability in the
future.

Packets now become parameterized over the backing buffer type, rather than
the lifetime of the buffer reference.  Backing buffers must meet various
requirements which are captured by the PacketBuffer trait.

Since almost all of our code relies on Packets being sourced from the buffer
stack and being of a specified size, rather than write Packet<Buffer<{
config::PACKET_BUFFER_SIZE }>> everywhere, I've also introduced the
BufferPacket alias, which is used through most of the packet processing
path.

Therefore most of the changes here are either:

* replacing Packet<'pktbuf> with Packet<impl PacketBuffer> in
  buffer-agnostic code

* replacing Packet<'pktbuf> with BufferPacket in code which is part of the
  packet pipeline

* deleting <'pktbuf> arguments in many places

The only real changes are in buffer_stack and the top of packet.